refactor(cli): keep watchAction flat, move only watchIgnore under watch/
watchAction stays at cli/actions/watchAction.ts (one file per action at the top
level); only the extracted helper lives in cli/actions/watch/watchIgnore.ts.
Split the test to mirror: watchAction.test.ts stays flat and the ignore-filter
cases move to watch/watchIgnore.test.ts.
